yoolo/yolov5/requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
onnx==1.8.1
onnxruntime
# scikit-learn==0.19.2  # for coreml quantization
onnx-simplifier    # for pytorch opset 11 constant folding issue

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP
